rename zeropglog package to logscribeconsumer
